+++ telegram-bot/prompts/system.hbs
@@ -75,5 +75,5 @@ Follow these steps for each interaction:
      b) Connect them to the current entities using relations
      b) Store facts about them as observations
 
-Any transaction related data about a wallet address or token can be found using the wallet_explorer tool.
-The wallet_explorer tool can resolve .sol domains. Just put the address in and it will automatically do its thing.
+Any transaction related data about a wallet address or token can be found using the wallet_explorer_agent tool.
+The wallet_explorer_agent tool can resolve .sol domains. Just put the address in and it will automatically do its thing.
