+++ telegram-bot/tools/raphgpt.ts
@@ -1,22 +1,12 @@
-import { openai } from "@ai-sdk/openai";
-import * as Bonfida from "@bonfida/spl-name-service";
 import { fmt } from "@grammyjs/parse-mode";
-import {
-  clusterApiUrl,
-  Connection,
-  LAMPORTS_PER_SOL,
-  PublicKey,
-} from "@solana/web3.js";
-import { generateText, tool } from "ai";
+import { tool } from "ai";
 import { browser } from "bot/browser";
 import logger from "bot/logger";
 import { telegram } from "bot/telegram";
 import type { ToolData } from "bot/tool-data";
-import { inspect } from "bun";
-import { format } from "date-fns";
-import { encoding_for_model } from "tiktoken";
 import { getEnv } from "utils/env";
 import { convertHtmlToMarkdown } from "utils/markdown";
+import { encoding_for_model } from "tiktoken";
 import { z } from "zod";
 
 /**
@@ -199,321 +189,6 @@ export function raphgptTools(data: ToolData) {
       },
     }),
 
-    /**
-     * Get real-time bus arrival information for a Singapore bus stop
-     * using the LTA DataMall v3 BusArrival API. Responses are updated every 20 seconds.
-     */
-    get_bus_arrival_timings: tool({
-      description:
-        "Get real-time bus arrival information for a Singapore bus stop using LTA DataMall v3 BusArrival API. " +
-        "Input must include a bus stop code and optional service number. Responses are updated every 20 seconds.",
-      parameters: z.object({
-        stop_id: z.string().describe("Bus stop code (BusStopCode)"),
-        service_no: z
-          .string()
-          .optional()
-          .describe("Bus service number (ServiceNo), optional"),
-      }),
-      async execute({ stop_id, service_no }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const params = new URLSearchParams();
-        params.append("BusStopCode", stop_id);
-        if (service_no) params.append("ServiceNo", service_no);
-        const resp = await fetch(
-          `https://datamall2.mytransport.sg/ltaodataservice/v3/BusArrival?${params}`,
-          { headers: { AccountKey: apiKey } },
-        );
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch bus arrival: ${resp.status} ${txt}`);
-        }
-        return await resp.json();
-      },
-    }),
-
-    /**
-     * Find bus stops by name using LTA DataMall BusStops API.
-     * Filters stops whose Description or RoadName contains the query (case-insensitive).
-     * Returns matching stops with BusStopCode, RoadName, Description, Latitude, Longitude.
-     */
-    find_bus_stops_by_name: tool({
-      description:
-        "Find bus stops whose description or road name contains a search string. Returns matching stops with bus stop code, road name, description, latitude, and longitude.",
-      parameters: z.object({
-        query: z.string().describe("Search string to filter bus stops"),
-      }),
-      async execute({ query }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const resp = await fetch(
-          "https://datamall2.mytransport.sg/ltaodataservice/BusStops",
-          { headers: { AccountKey: apiKey } },
-        );
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch bus stops: ${resp.status} ${txt}`);
-        }
-        const body = await resp.json();
-        logger.debug(
-          body,
-          "LTA bus stop data from call: find_bus_stops_by_name",
-        );
-
-        const stops = Array.isArray(body.value) ? body.value : body;
-        const q = query.toLowerCase();
-        return stops
-          .filter(
-            (s: any) =>
-              s.Description.toLowerCase().includes(q) ||
-              s.RoadName.toLowerCase().includes(q),
-          )
-          .map((s: any) => ({
-            BusStopCode: s.BusStopCode,
-            RoadName: s.RoadName,
-            Description: s.Description,
-            Latitude: parseFloat(s.Latitude),
-            Longitude: parseFloat(s.Longitude),
-          }));
-      },
-    }),
-
-    /**
-     * Find bus stops near a geographic point using LTA DataMall BusStops API.
-     * Returns stops within the given radius (km), sorted by distance ascending.
-     */
-    find_bus_stops_near_location: tool({
-      description:
-        "Find bus stops within a radius (km) of a given latitude/longitude. Returns stops with bus stop code, road name, description, latitude, longitude, and distance_km sorted by nearest first.",
-      parameters: z.object({
-        lat: z.number().describe("Latitude of the center point"),
-        lon: z.number().describe("Longitude of the center point"),
-        radius: z
-          .number()
-          .optional()
-          .describe("Search radius in kilometers (default 0.5)"),
-      }),
-      async execute({ lat, lon, radius = 0.5 }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const resp = await fetch(
-          "https://datamall2.mytransport.sg/ltaodataservice/BusStops",
-          { headers: { AccountKey: apiKey } },
-        );
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch bus stops: ${resp.status} ${txt}`);
-        }
-        const body = await resp.json();
-        logger.debug(
-          body,
-          "LTA bus stop data from call: find_bus_stops_near_location",
-        );
-
-        const stops = Array.isArray(body.value) ? body.value : body;
-        const toRad = (deg: number) => (deg * Math.PI) / 180;
-        const haversine = (
-          lat1: number,
-          lon1: number,
-          lat2: number,
-          lon2: number,
-        ) => {
-          const R = 6371; // km
-          const dLat = toRad(lat2 - lat1);
-          const dLon = toRad(lon2 - lon1);
-          const a =
-            Math.sin(dLat / 2) ** 2 +
-            Math.cos(toRad(lat1)) *
-              Math.cos(toRad(lat2)) *
-              Math.sin(dLon / 2) ** 2;
-          return R * (2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a)));
-        };
-        const results = stops
-          .map((s: any) => {
-            const lat2 = parseFloat(s.Latitude);
-            const lon2 = parseFloat(s.Longitude);
-            const distance_km = haversine(lat, lon, lat2, lon2);
-            return {
-              BusStopCode: s.BusStopCode,
-              RoadName: s.RoadName,
-              Description: s.Description,
-              Latitude: lat2,
-              Longitude: lon2,
-              distance_km,
-            };
-          })
-          .filter((s: any) => s.distance_km <= radius)
-          .sort((a: any, b: any) => a.distance_km - b.distance_km);
-        return results;
-      },
-    }),
-
-    /**
-     * Get detailed service information for all bus services in operation.
-     * Includes frequency, first/last stops, operator, etc.
-     */
-    get_bus_services: tool({
-      description:
-        "Get detailed service information for all bus services in operation, including first/last stops, operator codes, direction, category, and peak/off-peak frequencies.",
-      parameters: z.object({}),
-      async execute() {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const resp = await fetch(
-          "https://datamall2.mytransport.sg/ltaodataservice/BusServices",
-          { headers: { AccountKey: apiKey } },
-        );
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(
-            `Failed to fetch bus services: ${resp.status} ${txt}`,
-          );
-        }
-        const body = await resp.json();
-        logger.debug(body, "LTA bus stop data from call: get_bus_services");
-        return Array.isArray(body.value) ? body.value : body;
-      },
-    }),
-
-    /**
-     * Get detailed route information for all bus services in operation.
-     * Includes stop sequence, distances, and first/last bus timings per stop.
-     */
-    get_bus_routes: tool({
-      description:
-        "Get detailed route information for all bus services in operation, including stop sequence, distances, and first/last bus timings for weekdays, Saturdays and Sundays.",
-      parameters: z.object({}),
-      async execute() {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const resp = await fetch(
-          "https://datamall2.mytransport.sg/ltaodataservice/BusRoutes",
-          { headers: { AccountKey: apiKey } },
-        );
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch bus routes: ${resp.status} ${txt}`);
-        }
-        const body = await resp.json();
-        logger.debug(body, "LTA bus stop data from call: get_bus_routes");
-        return Array.isArray(body.value) ? body.value : body;
-      },
-    }),
-
-    /**
-     * Fetch passenger volume by bus stop from LTA DataMall PV/Bus.
-     * Optionally filter by date (YYYYMM).
-     */
-    get_passenger_volume_by_bus_stop: tool({
-      description:
-        "Get tap-in/tap-out passenger volume by bus stop. Optionally specify Date=YYYYMM to filter.",
-      parameters: z.object({
-        date: z
-          .string()
-          .optional()
-          .describe("YYYYMM for passenger volume data (optional)"),
-      }),
-      async execute({ date }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const params = new URLSearchParams();
-        if (date) params.append("Date", date);
-        const url =
-          "https://datamall2.mytransport.sg/ltaodataservice/PV/Bus" +
-          (params.toString() ? `?${params}` : "");
-        const resp = await fetch(url, { headers: { AccountKey: apiKey } });
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(
-            `Failed to fetch passenger volume PV/Bus: ${resp.status} ${txt}`,
-          );
-        }
-        return await resp.json();
-      },
-    }),
-
-    /**
-     * Fetch passenger volume by origin-destination for bus stops (PV/ODBus).
-     * Optionally filter by date (YYYYMM).
-     */
-    get_passenger_volume_by_od_bus: tool({
-      description:
-        "Get number of trips by weekdays and weekends from origin to destination bus stops. Optionally specify Date=YYYYMM.",
-      parameters: z.object({
-        date: z
-          .string()
-          .optional()
-          .describe("YYYYMM for origin-destination bus volume data (optional)"),
-      }),
-      async execute({ date }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const params = new URLSearchParams();
-        if (date) params.append("Date", date);
-        const url =
-          "https://datamall2.mytransport.sg/ltaodataservice/PV/ODBus" +
-          (params.toString() ? `?${params}` : "");
-        const resp = await fetch(url, { headers: { AccountKey: apiKey } });
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch PV/ODBus: ${resp.status} ${txt}`);
-        }
-        return await resp.json();
-      },
-    }),
-
-    /**
-     * Fetch passenger volume by origin-destination for train stations (PV/ODTrain).
-     * Optionally filter by date (YYYYMM).
-     */
-    get_passenger_volume_by_od_train: tool({
-      description:
-        "Get number of trips by weekdays and weekends from origin to destination train stations. Optionally specify Date=YYYYMM.",
-      parameters: z.object({
-        date: z
-          .string()
-          .optional()
-          .describe(
-            "YYYYMM for origin-destination train volume data (optional)",
-          ),
-      }),
-      async execute({ date }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const params = new URLSearchParams();
-        if (date) params.append("Date", date);
-        const url =
-          "https://datamall2.mytransport.sg/ltaodataservice/PV/ODTrain" +
-          (params.toString() ? `?${params}` : "");
-        const resp = await fetch(url, { headers: { AccountKey: apiKey } });
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch PV/ODTrain: ${resp.status} ${txt}`);
-        }
-        return await resp.json();
-      },
-    }),
-
-    /**
-     * Fetch passenger volume by train station (PV/Train).
-     * Optionally filter by date (YYYYMM).
-     */
-    get_passenger_volume_by_train: tool({
-      description:
-        "Get tap-in/tap-out passenger volume by train station. Optionally specify Date=YYYYMM.",
-      parameters: z.object({
-        date: z
-          .string()
-          .optional()
-          .describe("YYYYMM for train station volume data (optional)"),
-      }),
-      async execute({ date }) {
-        const apiKey = getEnv("LTA_DATAMALL_API_KEY");
-        const params = new URLSearchParams();
-        if (date) params.append("Date", date);
-        const url =
-          "https://datamall2.mytransport.sg/ltaodataservice/PV/Train" +
-          (params.toString() ? `?${params}` : "");
-        const resp = await fetch(url, { headers: { AccountKey: apiKey } });
-        if (!resp.ok) {
-          const txt = await resp.text();
-          throw new Error(`Failed to fetch PV/Train: ${resp.status} ${txt}`);
-        }
-        return await resp.json();
-      },
-    }),
 
     publish_mdx: tool({
       description: "Publish a webpage with MDX content",
@@ -572,146 +247,5 @@ ${url}`;
       },
     }),
 
-    wallet_explorer: tool({
-      description:
-        "Explore the Solana blockchain, using wallet addresses or transaction signatures",
-      parameters: z.object({
-        walletAddressOrSignature: z
-          .string()
-          .describe("Signature or address in base58 or .sol domain"),
-        instruction: z
-          .string()
-          .describe(
-            "Natural language instruction describing what you want from the address or signature",
-          ),
-      }),
-      async execute({ walletAddressOrSignature, instruction }) {
-        const connection = new Connection(clusterApiUrl("mainnet-beta"));
-
-        const { text: result, steps } = await generateText({
-          model: openai("o4-mini", {
-            structuredOutputs: false,
-            reasoningEffort: "high",
-          }),
-          system: `You are a Solana blockchain investigator.
-Current time in UTC: ${format(new Date(), "EEEE, yyyy-MM-dd 'at' HH:mm:ss zzz (XXX)")}.
-Always use get_sol_signatures before assuming there are no transactions associated with a specific wallet.`,
-          messages: [
-            {
-              role: "user",
-              content: [
-                {
-                  type: "text",
-                  text: instruction,
-                },
-                {
-                  type: "text",
-                  text: `Use the following parameter: ${walletAddressOrSignature}`,
-                },
-              ],
-            },
-          ],
-          tools: {
-            resolve_sol_domain: tool({
-              parameters: z.object({
-                domain: z.string().describe("Bonfida domain ending in .sol"),
-              }),
-              async execute({ domain }) {
-                const owner = await Bonfida.resolve(connection, domain);
-                return owner.toBase58();
-              },
-            }),
-
-            get_account_info: tool({
-              parameters: z.object({
-                walletAddress: z.string(),
-              }),
-              async execute({ walletAddress }) {
-                return await connection.getAccountInfoAndContext(
-                  new PublicKey(walletAddress),
-                );
-              },
-            }),
-
-            get_sol_signatures: tool({
-              description: "Get all signatures for wallet address",
-              parameters: z.object({
-                walletAddress: z.string(),
-                limit: z.number().optional(),
-              }),
-              async execute({ walletAddress, limit }) {
-                const signatures = await connection.getSignaturesForAddress(
-                  new PublicKey(walletAddress),
-                  {
-                    limit,
-                  },
-                );
-
-                logger.debug(signatures, "Confirmed signatures");
-
-                const formatted: string[] = [];
-
-                for (const sig of signatures) {
-                  const txDetails: string[] = [];
-
-                  txDetails.push(
-                    `timestamp: ${format(new Date(), "EEEE, yyyy-MM-dd 'at' HH:mm:ss zzz (XXX)")}`,
-                  );
-                  txDetails.push(`sig: ${sig.signature}`);
-                  txDetails.push(`memo: ${sig.memo}`);
-                  txDetails.push(`error: ${inspect(sig.err)}`);
-
-                  formatted.push(txDetails.join(","));
-                }
-
-                return formatted.join("\n");
-              },
-            }),
-
-            get_sol_tx: tool({
-              description: "Get transaction by signature",
-              parameters: z.object({
-                sig: z.string(),
-              }),
-              async execute({ sig }) {
-                const transaction = await connection.getParsedTransaction(sig, {
-                  commitment: "confirmed",
-                  maxSupportedTransactionVersion: 0,
-                });
-                if (!transaction) throw new Error("Transaction not found");
-
-                const formatted: string[] = [];
-                const txDetails: string[] = [];
-
-                txDetails.push(
-                  `timestamp: ${format(new Date(), "EEEE, yyyy-MM-dd 'at' HH:mm:ss zzz (XXX)")}`,
-                );
-                txDetails.push(`data: ${inspect(transaction.meta)}`);
-
-                formatted.push(txDetails.join(","));
-
-                return formatted.join("\n");
-              },
-            }),
-
-            lamports_to_sol: tool({
-              description: "Calculate lamports to sol",
-              parameters: z.object({
-                lamports: z.number(),
-              }),
-              async execute({ lamports }) {
-                return lamports / LAMPORTS_PER_SOL;
-              },
-            }),
-          },
-          maxSteps: 5,
-        });
-
-        logger.debug(steps, "Wallet explorer resulting steps");
-        logger.debug(result);
-
-        return `Result: ${result}`;
-      },
-    }),
   };
 }
